+++ kong/tools/mime_type.lua
@@ -3,6 +3,10 @@ local lpeg = require "lpeg"
 local P, S, R, C = lpeg.P, lpeg.S, lpeg.R, lpeg.C
 local ipairs = ipairs
 local lower = string.lower
+local sub = string.sub
+local find = string.find
+
+local WILDCARD = "*"
 
 --[[
 RFC2045(https://www.ietf.org/rfc/rfc2045.txt)
@@ -88,7 +92,44 @@ local function parse_mime_type(mime_type)
   return media_type:match(mime_type)
 end
 
+--- Checks if this mime-type includes other mime-type
+-- @tparam table this This mime-type
+-- @tparam table other Other mime-type
+-- @treturn boolean Returns `true` if this mime-type includes other, `false` otherwise
+local function includes(this, other)
+  if not other then
+    return false
+  end
+
+  local this_type = this.type
+  local this_subtype = this.subtype
+  local other_type = other.type
+  local other_subtype = other.subtype
+
+  if this_type == WILDCARD then
+    -- */* includes anything
+    return true
+  end
+
+  if this_type == other_type then
+    if this_subtype == other_subtype or this_subtype == WILDCARD then
+      return true
+    end
+
+    if sub(this_subtype, 1, 2) == "*+" then
+      -- subtype is wildcard with suffix, e.g. *+json
+      local this_subtype_suffix = sub(this.subtype, 3)
+      local pattern = "^.-+" .. this_subtype_suffix ..  "$"
+      if find(other_subtype, pattern) then
+        return true
+      end
+    end
+  end
+
+  return false
+end
 
 return {
-  parse_mime_type = parse_mime_type
+  parse_mime_type = parse_mime_type,
+  includes = includes,
 }